% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbExtract.R
\name{ExtractVarTmb}
\alias{ExtractVarTmb}
\title{Function to extract  variances parameters from a list containing TMB::MakeADFun and nlminb objects}
\usage{
ExtractVarTmb(obj, params, reNames = NULL)
}
\arguments{
\item{obj}{A list that contains the TMB::MakeADFun and nlminb objects.}

\item{params}{Parameter names to extract. If NULL, all parameters will be extracted.}

\item{reNames}{A vector of names to rename parameters. If NULL, the original TMB names will be retained.}
}
\value{
A vector of chosen parameters.
}
\description{
Function to extract  variances parameters from a list containing TMB::MakeADFun and nlminb objects
}
\author{
Jemay Salomon
}
