% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbExtract-package.R
\docType{package}
\name{tmbExtract-package}
\alias{tmbExtract-package}
\title{tmbExtract: Extract TMB parameters}
\description{
Extract TMB parameters from a list containing TMB::MakeADFun and nlminb objects.
}
\author{
\strong{Maintainer}: Jemay Salomon \email{jemay.salomon@inrae.fr} (\href{https://orcid.org/0000-0003-4454-7916}{ORCID})

}
\keyword{internal}
