% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.R
\name{tmbExtract}
\alias{tmbExtract}
\title{tmbExtract}
\usage{
tmbExtract(tmbObj, params = NULL, reNames = NULL, paramsType)
}
\arguments{
\item{tmbObj}{A list containing TMB::MakeADFun and nlminb tmbObjects.}

\item{params}{Parameter names to extract. If NULL, all parameters will be extracted.}

\item{reNames}{A vector of names to rename parameters. If NULL, the original TMB names will be retained.}

\item{paramsType}{Specifies the type of TMB parameters to extract (e.g., "paramsTmb", "random", "variance", "correlation").}
}
\value{
A vector of selected parameters.
}
\description{
Macro function to extract TMB parameters of specified types.
}
\author{
Jemay Salomon
}
