% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TMBusers.version}
\alias{TMBusers.version}
\title{Get the Version of the TMBusers Package}
\usage{
TMBusers.version()
}
\value{
A list containing the package version.
}
\description{
This function returns the version of the \code{TMBusers} package.
}
\examples{
TMBusers.version()
}
\author{
Jemay Salomon
}
