% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract.R
\name{ExtractVarTmb}
\alias{ExtractVarTmb}
\title{ExtractVarTmb}
\usage{
ExtractVarTmb(tmbObj, params, reNames = NULL)
}
\arguments{
\item{tmbObj}{A list that contains the TMB::MakeADFun and nlminb tmbObjects.}

\item{params}{Parameter names to extract. If NULL, all parameters will be extracted.}

\item{reNames}{A vector of names to rename parameters. If NULL, the original TMB names will be retained.}
}
\value{
A vector of chosen parameters.
}
\description{
Function to extract  variances parameters from a list containing TMB::MakeADFun and nlminb Objects
}
\author{
Jemay Salomon
}
